curso1/requirements.txt
openai==1.66.2
langchain==0.2.0
langchain-openai
python-dotenv==1.0.1
faiss-cpu==1.10.0
langchain-community
pypdf==5.3.1
